+++ files/index.d.ts
@@ -548,6 +548,29 @@ Notes:
 export function find<T>(predicate: (x: T) => boolean): (list: T[]) => T | undefined;
 
 /*
+Method: exists
+
+Explanation: It returns `true` if there is at least one element in `list` that satisfy the `predicate`.
+
+Example:
+
+```
+const predicate = x => R.type(x.foo) === 'Number'
+const list = [{foo: 'bar'}, {foo: 1}]
+
+const result = R.exists(predicate)(list)
+// => true
+```
+
+Categories: List
+
+Notes:
+
+*/
+// @SINGLE_MARKER
+export function find<T>(predicate: (x: T) => boolean): (list: T[]) => T | undefined;
+
+/*
 Method: findNth
 
 Explanation: It returns the `nth` element of `list` that satisfy the `predicate` function.
@@ -754,8 +777,8 @@ Example:
 
 ```
 const result = [
-  R.includes('oo')('foo'),
-  R.includes({a: 1})([{a: 1}])
+  R.includes('foo')('oo'),
+  R.includes([{a: 1}])({a: 1})
 ]
 // => [true, true ]
 ```
@@ -766,8 +789,8 @@ Notes:
 
 */
 // @SINGLE_MARKER
-export function includes(s: string): (list: readonly string[] | string) => boolean;
-export function includes<T>(target: T): (list: readonly T[]) => boolean;
+export function includes(list: readonly string[] | string): (substringToFind: string) => boolean;
+export function includes<T>(list: readonly T[]): (target: T) => boolean;
 
 /*
 Method: excludes
@@ -780,8 +803,8 @@ Example:
 
 ```
 const result = [
-  R.excludes('ar')('foo'),
-  R.excludes({a: 2})([{a: 1}])
+  R.excludes('foo')('ar'),
+  R.excludes([{a: 1}])({a: 2})
 ]
 // => [true, true ]
 ```
@@ -792,8 +815,8 @@ Notes:
 
 */
 // @SINGLE_MARKER
-export function excludes<T extends string>(valueToFind: T): (input: string) => boolean;
-export function excludes<T>(valueToFind: T): (input: T[]) => boolean;
+export function excludes(list: readonly string[] | string): (substringToFind: string) => boolean;
+export function excludes<T>(list: readonly T[]): (target: T) => boolean;
 
 /*
 Method: indexOf
@@ -859,7 +882,7 @@ const result = R.intersection(listA)(listB)
 
 Categories: List
 
-Notes: There is slight difference between Rambda and Ramda implementation. Ramda.intersection(['a', 'b', 'c'], ['c', 'b']) result is "[ 'c', 'b' ]", but Rambda result is "[ 'b', 'c' ]".
+Notes:
 
 */
 // @SINGLE_MARKER
@@ -3000,9 +3023,34 @@ Notes:
 export function splitEvery<T>(sliceLength: number): (input: T[]) => (T[])[];
 
 /*
-Method: symmetricDifference
+Method: difference
 
 Explanation: It returns a merged list of `x` and `y` with all equal elements removed.
+
+`R.equals` is used to determine equality.
+
+Example:
+
+```
+const x = [ 1, 2, 3, 4 ]
+const y = [ 3, 4, 5, 6 ]
+
+const result = R.difference(x)(y)
+// => [ 1, 2, 5, 6 ]
+```
+
+Categories: List
+
+Notes:
+
+*/
+// @SINGLE_MARKER
+export function difference<T>(x: T[]): (y: T[]) => T[];
+
+/*
+Method: symmetricDifference
+
+Explanation: It returns all items that are in either of the lists, but not in both.
 
 `R.equals` is used to determine equality.
 
@@ -3022,7 +3070,7 @@ Notes:
 
 */
 // @SINGLE_MARKER
-export function symmetricDifference<T>(x: T[]): <T>(y: T[]) => T[];
+export function symmetricDifference<T>(list: T[]): (list: T[]) => T[];
 
 /*
 Method: tail
@@ -4140,14 +4188,13 @@ export function split(separator: string | RegExp): (str: string) => string[];
 /*
 Method: range
 
-Explanation: It returns list of numbers between `startInclusive` to `endExclusive` markers.
-If `start` is greater than `end`, then the result will be in descending order.
+Explanation: It returns list of numbers between `startInclusive` to `endInclusive` markers.
 
 Example:
 
 ```
-[R.range(0)(5), R.range(5)(0)]
-// => [[0, 1, 2, 3, 4], [5, 4, 3, 2, 1]]
+[R.range(5), R.range(1, 5)]
+// => [[0, 1, 2, 3, 4, 5], [1, 2, 3, 4, 5]]
 ```
 
 Categories: Number
@@ -4156,7 +4203,25 @@ Notes:
 
 */
 // @SINGLE_MARKER
-export function range(startInclusive: number): (endExclusive: number) => number[];
+export function range(startInclusive: number, endInclusive: number) : number[];
+
+/*
+Method: rangeDescending
+
+Explanation: It returns list of numbers between `endInclusive` to `startInclusive` markers.
+
+Example:
+
+```
+```
+
+Categories: Number
+
+Notes:
+
+*/
+// @SINGLE_MARKER
+export function rangeDescending(startInclusive: number, endInclusive: number) : number[];
 
 /*
 Method: pipeAsync